Move tests for PanoptesPluginInfo into new test_base_plugin.py from test_framework.py to collocate with other tests for classes from panoptes_base_plugin.py.
